refactor: split plugin/cli code, use version from package.json
