+++ src/app/controller/checkForVariables.ts
