+++ src/app/controller/config.ts
@@ -1,5 +1,6 @@
 export const config = {
   frameWidth: 300,
   docsLink: "https://github.com/PavelLaptev/figma-variables-export",
-  changelogLink: "https://github.com/PavelLaptev/tokens-bruecke#changelog",
+  changelogLink:
+    "https://github.com/PavelLaptev/tokens-bruecke/blob/main/CHANGELOG.md",
 };
