+++ src/app/views/SettingsView/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from "react";
 import styles from "./styles.module.scss";
+import pkg from "../../../../package.json";
 
 import {
   Panel,
@@ -15,16 +16,16 @@ import {
   OverlayList,
 } from "pavelLaptev/react-figma-ui/ui";
 
-import { config } from "../../../utils/config";
+import { config } from "../../controller/config";
 
 import { Toast } from "../../components/Toast";
 import { ServerSettingsView } from "../ServerSettingsView";
 
-import { pushToJSONBin } from "../../../utils/servers/pushToJSONBin";
-import { pushToGithub } from "../../../utils/servers/pushToGithub";
-import { githubPullRequest } from "../../../utils/servers/githubPullRequest";
-import { pushToGitlab } from "../../../utils/servers/pushToGitlab";
-import { pushToCustomURL } from "../../../utils/servers/pushToCustomURL";
+import { pushToJSONBin } from "../../api/servers/pushToJSONBin";
+import { pushToGithub } from "../../api/servers/pushToGithub";
+import { githubPullRequest } from "../../api/servers/githubPullRequest";
+import { pushToGitlab } from "../../api/servers/pushToGitlab";
+import { pushToCustomURL } from "../../api/servers/pushToCustomURL";
 
 import { downloadTokensFile } from "../../api/downloadTokensFile";
 
@@ -46,6 +47,7 @@ interface ViewProps {
   currentView: string;
 }
 
+const version = pkg.version;
 const stylesList = [
   {
     id: "text",
@@ -672,7 +674,7 @@ export const SettingsView = (props: ViewProps) => {
               <Text>Documentation</Text>
             </a>
             <a href={config.changelogLink} target="_blank">
-              <Text>v.2.5.0</Text>
+              <Text>v.{version}</Text>
             </a>
           </Stack>
         </Panel>
