+++ src/common/transform/color/convertFigmaLinearGradient.ts
