+++ src/common/transform/countTokens.ts
