+++ src/common/transform/groupObjectNamesIntoCategories.ts
