+++ src/common/transform/removeDollarSign.ts
