+++ src/common/transform/text/getFontStyleAndWeight.ts
