+++ src/common/transform/text/getLetterSpacing.ts
