--- src/utils/normalizeValue.spec.ts
@@ -1,21 +1,0 @@
-import { describe, expect, test } from "vitest";
-import { IResolver } from "../common/resolver";
-
-import { normalizeValue } from "./normalizeValue";
-
-const resolver = {} as IResolver;
-
-describe("getFontStyleAndWeight", () => {
-  test("Lossy float", () => {
-    expect(
-      normalizeValue({
-        variableValue: 0.40000000200345043,
-        variableType: "FLOAT",
-        variableScope: [],
-        colorMode: "hex",
-        useDTCGKeys: false,
-        includeValueAliasString: false,
-      }, resolver)
-    ).toBe("0.4px")
-  });
-});
