+++ tsconfig.base.json
@@ -8,6 +8,7 @@
     "noImplicitAny": false,
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
+    "resolveJsonModule": true,
     "lib": [
       "es2021"
     ],
